Seperate API module from main file
