+++ open-messages.rb
@@ -16,7 +16,7 @@ require 'kramdown'
 require 'password_blacklist'
 require 'securerandom'
 require 'sequel'
-
+require_relative 'API'
 Cuba.plugin Cuba::Safe
 Cuba.plugin Cuba::Render
 Cuba.use Rack::Session::Cookie, secret: Random.new_seed.to_s
@@ -28,111 +28,6 @@ Cuba.use Rack::Static, root: 'public', urls: ['/js']
 Cuba.settings[:render][:template_engine] = 'html.erb'
 Cuba.settings[:render][:views] = './views'
 
-module API
-  DB = Sequel.connect('sqlite://db/sqlite.db')
-  data = DB[:data]
-  user = DB[:user]
-  class GetMessage < Cuba; end
-  GetMessage.define do
-    on ':title' do |title|
-      article = data.where(title: title).get(:textarea)
-      @markdown2html = Kramdown::Document.new(article).to_html
-      @html2markdown = Kramdown::Document.new(@markdown2html, input: 'html')
-                                         .to_kramdown
-    rescue NoMethodError
-      res.status = 404
-      res.redirect('/404')
-    else
-      res.write view('messages')
-    end
-  end
-  class GetAllUserMessages < Cuba; end
-  GetAllUserMessages.define do
-    on ':username' do |username|
-      @user_messages_title = data.where(username: username).select_map(:title)
-      if @user_messages_title.any?
-        @array = @user_messages_title.to_a
-        res.write view('getallusermessages')
-      elsif @user_messages_title.empty?
-        res.redirect('/404')
-      end
-    end
-  end
-  class EditMessage < Cuba; end
-  EditMessage.define do
-    on root, param('username'), param('password'), param('title'), param('textarea') do |username, password, title, textarea|
-      check_password = BCrypt::Password.new(user.where(username: username).get(:password)).is_password?(password)
-      if check_password == true
-        DB.transaction do
-          data.where(title: title.downcase.strip.tr(' ', '-').gsub(/[^\w-]/, '').to_s)
-              .update(textarea: textarea)
-          data.where(title: title.downcase.strip.tr(' ', '-').gsub(/[^\w-]/, '').to_s)
-              .update(edited_on: Date.today.to_s)
-        end
-        res.redirect("/message/get/#{title.downcase.strip.tr(' ', '-').gsub(/[^\w-]/, '')}")
-      elsif check_password == false
-        res.redirect('/put_error')
-      end
-    rescue BCrypt::Errors::InvalidHash
-      res.redirect('/put_error')
-    end
-  end
-  class DeleteMessage < Cuba; end
-  DeleteMessage.define do
-    on root, param('title'), param('username'), param('password') do |title, username, password|
-      check_password = BCrypt::Password.new(user.where(username: username).get(:password)).is_password?(password)
-      if check_password == true
-        data.where(title: title, username: username).delete
-        res.status = 200
-      elsif check_password == false
-        res.status = 404
-      end
-    rescue BCrypt::Error
-      res.status = 500
-    rescue Standard::Error
-      res.status = 404
-    end
-  end
-  class PutMessage < Cuba; end
-  PutMessage.define do
-    on root, param('username'), param('password'), param('title'), param('textarea') do |username, password, title, textarea|
-      generate_id = SecureRandom.uuid
-      begin
-        check_password = BCrypt::Password.new(user.where(username: username).get(:password)).is_password?(password)
-        if check_password == true
-          data.insert(uuid: generate_id.to_s, username: username.to_s,
-                      title: title, created_on: Date.today.to_s, textarea: textarea.to_s)
-          res.redirect("/message/get/#{title.downcase.strip.tr(' ', '-').gsub(/[^\w-]/, '')}")
-        elsif check_password == false
-          res.redirect('/put_error')
-        end
-      rescue BCrypt::Errors::InvalidHash
-        res.redirect('/put_error')
-      end
-    end
-  end
-
-  class SignUp < Cuba; end
-  SignUp.define do
-    on root, param('username'), param('password') do |username, password|
-      checker = PasswordBlacklist::Checker.new
-      if user.where(username: username).first
-        @used_username = '<small> * Username already in use </small>'
-        res.status = 500
-        res.write view('/signup')
-      elsif checker.blacklisted?(password) == true
-        @blacklist_password = '<small> * The password provided is blacklisted </small>'
-        res.status = 500
-        res.write view('/signup')
-      else
-        bcrypted_password = BCrypt::Password.create(password)
-        user.insert(username: username, password: bcrypted_password)
-        hit_status = res.status = 200
-        res.redirect('/') if hit_status
-      end
-    end
-  end
-end
 module FRONTEND
   class Root < Cuba; end
   Root.define do
@@ -202,6 +97,9 @@ Cuba.define do
     on 'message/user' do
       run API::GetAllUserMessages
     end
+    #on 'message/title' do
+    #  run API::GetAllTitleRevisions
+    #end
     on 'put_error' do
       run FRONTEND::FrontendPutError
     end
